Implement SecurityLevelChecker for SecureLogin and BrowserLock (#824)
The SecurityLevelChecker was accidentally removed in the PR(#818). This
patch reimplements the SecurityLevelChecker for SecureLogin and
BrowserLock in PrivacySettings.

Refs: #809
