+++ src/org/triple/banana/settings/ExtensionFeatures.java
@@ -18,8 +18,6 @@ import org.banana.cake.interfaces.BananaApplicationUtils;
 import org.banana.cake.interfaces.BananaFeatureFlags;
 import org.triple.banana.R;
 import org.triple.banana.appmenu.AppMenuDelegate;
-import org.triple.banana.authentication.SecurityLevelChecker;
-import org.triple.banana.authentication.SecurityLevelChecker.SecurityLevel;
 import org.triple.banana.toolbar.ToolbarEditor;
 
 public class ExtensionFeatures extends PreferenceFragmentCompat {
@@ -64,20 +62,6 @@ public class ExtensionFeatures extends PreferenceFragmentCompat {
         });
     }
 
-    @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        SecurityLevelChecker.get().addListener(this::onSecurityLevelChanged);
-    }
-
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
-        // FIXME(#313): Need a good way to prevent memory leak. Because 'onDestroy' is not
-        // guaranteed to be called in the fragment.
-        SecurityLevelChecker.get().removeListener(this::onSecurityLevelChanged);
-    }
-
     private void showRestartDialog() {
         AlertDialog.Builder builder = BananaApplicationUtils.get().getDialogBuilder(getActivity());
         builder.setMessage(R.string.restart_message)
@@ -86,10 +70,6 @@ public class ExtensionFeatures extends PreferenceFragmentCompat {
                         (dialog, which) -> { BananaApplicationUtils.get().restart(); })
                 .create()
                 .show();
-    }
-
-    public void onSecurityLevelChanged(SecurityLevel newLevel) {
-        boolean isSecure = newLevel == SecurityLevel.SECURE;
     }
 
     public static class FeatureName {
